Add a bunch more synch problems solutions from the archive.
Some of them (mostly the older ones) don't have the problem statement,
but you can probably figure out the general idea.

I haven't added them to the system (either the makefiles or the menu)
but you should be able to do that yourself easily enough at this
point. If in doubt, examine the airballoon commit in the history.